refactor(web-rdx): add react router dom and refactor app folder
