+++ packages/webapp-redux/package.json
@@ -11,6 +11,7 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-redux": "7.2.4",
+    "react-router-dom": "5.2.0",
     "react-scripts": "4.0.3",
     "typescript": "4.3.2"
   },
@@ -22,6 +23,7 @@
     "@types/node": "15.12.2",
     "@types/react": "17.0.11",
     "@types/react-redux": "7.1.16",
+    "@types/react-router-dom": "5.1.7",
     "@types/react-dom": "17.0.7",
     "react-app-rewired": "2.1.8",
     "ts-jest": "26.5.6",
