+++ packages/webapp-redux/tsconfig.json
@@ -1,5 +1,7 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src"],
+  "include": [
+    "src"
+  ],
   "compilerOptions": {}
 }
